LayoutTests/imported/w3c/web-platform-tests/workers/name-property-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = Unhandled rejection: Can't find variable: SharedWorker

PASS name property value for DedicatedWorkerGlobalScope
PASS name property is replaceable for DedicatedWorkerGlobalScope
PASS Declaring name as an accidental global must not cause a harness error for DedicatedWorkerGlobalScope
